+++ cmake/CpackCtest.cmake
@@ -1,3 +1,4 @@
+
 #------------------------------------------------------------------------------
 # CPack Configurations
 #------------------------------------------------------------------------------
@@ -41,11 +42,26 @@ if (BUILD_TESTING)
   message(STATUS "Build test files ENABLED")
   ENABLE_TESTING()
   include(CTest)
+  if (CMAKE_CXX_COMPILER_ID STREQUAL "Clang"  OR CMAKE_CXX_COMPILER_ID STREQUAL "AppleClang")
+    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g -O0 -Wall -Wextra\
+       -pedantic -Wno-unknown-pragmas -W -Wshadow -Wunused-variable  \
+       -Wunused-parameter -Wunused-function   -Werror=type-limits -Wno-unneeded-internal-declaration -Wno-unused-function  -Wno-delete-non-abstract-non-virtual-dtor   -Wunused  -Wno-long-long\
+        -Wno-system-headers -Wno-deprecated -Wno-dtor-name -Woverloaded-virtual -Wwrite-strings")
+
+    endif()
   if (CMAKE_COMPILER_IS_GNUCXX)
-    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g -O0 -Wall -Wextra -pedantic -W -Wshadow -Wunused-variable  -Wunused-parameter -Wunused-function        -Wunused  -Wno-long-long -Wno-system-headers -Wno-deprecated -Woverloaded-virtual -Wwrite-strings -fprofile-arcs -ftest-coverage")
+    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g -O0 -Wall -Wextra\
+         -pedantic -Wno-unknown-pragmas -W -Wshadow -Wunused-variable  \
+         -Wunused-parameter -Wunused-function  -Werror=type-limits  -Wno-unneeded-internal-declaration -Wno-unused-function  -Wno-delete-non-abstract-non-virtual-dtor   -Wunused  -Wno-long-long\
+          -Wno-system-headers -Wno-deprecated -Wno-dtor-name -Woverloaded-virtual -Wwrite-strings")
   endif()
   if (CMAKE_COMPILER_IS_GNUCC)
-    set(CMAKE_C_FLAGS_DEBUG "${CMAKE_C_FLAGS_DEBUG} -g -O0 -Wall -Wextra -W -Wno-long-long -pedantic -fprofile-arcs -ftest-coverage")
+    set(CMAKE_C_FLAGS_DEBUG "${CMAKE_C_FLAGS_DEBUG} -g -O0 -Wall -Wextra -W  -Wno-delete-non-abstract-non-virtual-dtor -Wno-unknown-pragmas\
+        -Wno-long-long -pedantic")
+  endif()
+  if (CMAKE_C_COMPILER_ID STREQUAL "Clang")
+    set(CMAKE_C_FLAGS_DEBUG "${CMAKE_C_FLAGS_DEBUG} -g -O0 -Wall -Wextra -W  -Wno-delete-non-abstract-non-virtual-dtor -Wno-unknown-pragmas\
+      -Wno-long-long -pedantic")
   endif()
   add_subdirectory (${PROJECT_SOURCE_DIR}/tests)
 else()