+++ tests/geometry/curves/estimation/testEstimatorComparator.cpp
@@ -74,7 +74,6 @@ bool testCompareEstimator(const std::string &name, const Shape & aShape, double 
            + name ). c_str() );
   
   // Creates a digitizer on the window (xLow, xUp).
-  typedef Space::RealPoint RealPoint;
   RealPoint xLow( -10.0, -10.0 );
   RealPoint xUp( 10.0, 10.0 );
   GaussDigitizer<Space,Shape> dig;