+++ tests/shapes/testShapeMoveCenter.cpp
@@ -51,9 +51,6 @@
 using namespace DGtal;
 using namespace Z2i;
 
-typedef Z2i::Space     Space;
-typedef Z2i::Point     Point;
-typedef Z2i::RealPoint RealPoint;
 
 typedef AccFlower2D<Space>  MyAccFlower;
 typedef Astroid2D<Space>    MyAstroid;