+++ ChangeLog.md
@@ -1,9 +1,9 @@
 # DGtal 1.4beta
 
 ## New features / critical changes
- 
+
 - *General*
-  - Major update: C++17 is now required for DGtal. (David Coeurjolly, 
+  - Major update: C++17 is now required for DGtal. (David Coeurjolly,
     [#1682](https://github.com/DGtal-team/DGtal/pull/1682))
 
 ## Changes
@@ -27,7 +27,11 @@
   - Removing unnecessary includes to speed-up compilation (David Coeurjolly,
     [#1680](https://github.com/DGtal-team/DGtal/pull/1680))
   - Upgrading pybind11 to v2.9 or python binding (David Coeurjolly,
-    [#1685](https://github.com/DGtal-team/DGtal/pull/1685)) 
+    [#1685](https://github.com/DGtal-team/DGtal/pull/1685))
+  - Many warning fixed (due to c++17 upgrade. (David Coeurjolly,
+    [#1691](https://github.com/DGtal-team/DGtal/pull/1691))
+  - WITH_COVERAGE option removed. (David Coeurjolly,
+    [#1691](https://github.com/DGtal-team/DGtal/pull/1691))
 
 - *Topology package*
   - Fix KhalimskySpaceND to get it work with BigInteger (Tristan Roussillon,